+++ enterprise/cmd/frontend/auth/httpheader/middleware_test.go
@@ -14,13 +14,15 @@ import (
 	"github.com/sourcegraph/sourcegraph/pkg/actor"
 	"github.com/sourcegraph/sourcegraph/pkg/conf"
 	"github.com/sourcegraph/sourcegraph/schema"
+
+	"golang.org/x/crypto/ssh"
 )
 
 // SEE ALSO FOR MANUAL TESTING: See the Middleware docstring for information about the testproxy
 // helper program, which helps with manual testing of the HTTP auth proxy behavior.
 func TestMiddleware(t *testing.T) {
-	licensing.MockGetConfiguredProductLicenseInfo = func() (*license.Info, error) {
-		return &license.Info{Tags: licensing.EnterpriseTags}, nil
+	licensing.MockGetConfiguredProductLicenseInfo = func() (*license.Info, *ssh.Signature, error) {
+		return &license.Info{Tags: licensing.EnterpriseTags}, nil, nil
 	}
 	defer func() { licensing.MockGetConfiguredProductLicenseInfo = nil }()