+++ enterprise/cmd/frontend/internal/licensing/enforcement_test.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/sourcegraph/enterprise/pkg/license"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/db"
+	"golang.org/x/crypto/ssh"
 )
 
 func TestEnforcementPreCreateUser(t *testing.T) {
@@ -68,8 +69,8 @@ func TestEnforcementPreCreateUser(t *testing.T) {
 	}
 	for _, test := range tests {
 		t.Run(fmt.Sprintf("license %s with %d active users", test.license, test.activeUserCount), func(t *testing.T) {
-			MockGetConfiguredProductLicenseInfo = func() (*license.Info, error) {
-				return test.license, nil
+			MockGetConfiguredProductLicenseInfo = func() (*license.Info, *ssh.Signature, error) {
+				return test.license, nil, nil
 			}
 			defer func() { MockGetConfiguredProductLicenseInfo = nil }()
 			db.Mocks.Users.Count = func(context.Context, *db.UsersListOptions) (int, error) {